Move TestGet[Tenant|User]FromEchoContext to util package
